refactor: move types to designated folder
