+++ react-ts-app/src/pages/HomePage.tsx
@@ -1,15 +1,7 @@
 import React from "react";
 import { Cards } from "../components/Cards/Cards";
 import { SearchElem } from "../components/Search/Search";
-
-export interface ISearchResult {
-  id: string;
-  title: string;
-  imageUrl: string;
-  date: Date;
-  author: string;
-  location: string;
-}
+import { ISearchResult } from "../types/types";
 
 export const Home = () => {
   const [searchResults, setSearchResults] = React.useState<ISearchResult[]>([]);
